requirements.txt
# Основной фреймворк
Django>=4.2,<5.0                 # Django до 5.0 (для совместимости с celery-beat)
djangorestframework>=3.14,<3.15 # DRF
djangorestframework-simplejwt>=5.2,<5.3
django-cors-headers>=4.2,<4.3

# Работа с медиафайлами и статикой
Pillow>=10.0,<10.1
whitenoise>=6.5,<6.6

# Документация и фильтрация
drf-yasg>=1.21,<1.22
django-filter>=24.3

# Асинхронные задачи и уведомления
firebase-admin==6.5.0
celery>=5.5,<5.6
django-celery-beat>=2.5,<2.6

# Развёртывание
gunicorn>=21.0,<22.0
daphne>=4.0,<4.1  # ASGI-сервер для WebSocket

# Кэширование и базы данных
redis>=5.0,<5.1
psycopg2-binary>=2.9,<2.10

# Дополнительные утилиты
python-decouple>=3.8,<3.9
requests>=2.31,<2.32

# WebSocket
channels>=4.0,<4.1
channels-redis>=4.2,<4.3
